+++ src/platform/canvas/PathCairo.h
@@ -41,19 +41,19 @@ public:
 
     // For CanvasPath
     virtual void closePath() override;
-    virtual void moveTo(double x, double y) override;
-    virtual void lineTo(double x, double y) override;
-    virtual void quadraticCurveTo(double cpx, double cpy, double x,
-                                  double y) override;
-    virtual void bezierCurveTo(double cp1x, double cp1y, double cp2x,
-                               double cp2y, double x, double y) override;
-    virtual void arcTo(double x1, double y1, double x2, double y2,
-                       double radius) override;
-    virtual void rect(double x, double y, double w, double h) override;
-    virtual void arc(double x, double y, double radius, double startAngle,
-                     double endAngle, bool anticlockwise = false) override;
-    virtual void ellipse(double x, double y, double radiusX, double radiusY,
-                         double rotation, double startAngle, double endAngle,
+    virtual void moveTo(float x, float y) override;
+    virtual void lineTo(float x, float y) override;
+    virtual void quadraticCurveTo(float cpx, float cpy, float x,
+                                  float y) override;
+    virtual void bezierCurveTo(float cp1x, float cp1y, float cp2x, float cp2y,
+                               float x, float y) override;
+    virtual void arcTo(float x1, float y1, float x2, float y2,
+                       float radius) override;
+    virtual void rect(float x, float y, float w, float h) override;
+    virtual void arc(float x, float y, float radius, float startAngle,
+                     float endAngle, bool anticlockwise = false) override;
+    virtual void ellipse(float x, float y, float radiusX, float radiusY,
+                         float rotation, float startAngle, float endAngle,
                          bool anticlockwise = false) override;
 
 private:
